refactor(webview): extract KeyManagementHandler from webview-provider
